FIX: Allow more than 65536 sequences in fasta
fasta's made from VCFs typically have many more than 65536 sequences,
but the indexing method only stores them using 16-bit integers.

Increase the number number of sequences that can be stored by using
32-bit integers for indexing.

+++ VARSCOT_pipeline/read_mapping/common.h
@@ -9,7 +9,7 @@ namespace seqan {
 template <typename TChar, typename TOwner>
 struct SAValue<StringSet<String<TChar>, TOwner > >
 {
-    typedef Pair<uint16_t, uint32_t> Type;
+    typedef Pair<uint32_t, uint32_t> Type;
 };
 
 template <typename TChar, typename TOwner>
